refactor(thinking): extracted all state manipulations to useThinking hook
